+++ net/minecraft/fluid/WaterFluid.java
@@ -56,7 +56,7 @@ extends BaseFluid {
                 world.playSound((double)blockPos.getX() + 0.5, (double)blockPos.getY() + 0.5, (double)blockPos.getZ() + 0.5, SoundEvents.BLOCK_WATER_AMBIENT, SoundCategory.BLOCKS, random.nextFloat() * 0.25f + 0.75f, random.nextFloat() + 0.5f, false);
             }
         } else if (random.nextInt(10) == 0) {
-            world.addParticle(ParticleTypes.UNDERWATER, (float)blockPos.getX() + random.nextFloat(), (float)blockPos.getY() + random.nextFloat(), (float)blockPos.getZ() + random.nextFloat(), 0.0, 0.0, 0.0);
+            world.addParticle(ParticleTypes.UNDERWATER, (double)blockPos.getX() + (double)random.nextFloat(), (double)blockPos.getY() + (double)random.nextFloat(), (double)blockPos.getZ() + (double)random.nextFloat(), 0.0, 0.0, 0.0);
         }
     }
